Logout and role-based-access

See merge request informatik/qe/swess23/group2/g2t4!24

+++ frontend/src/services/auth.service.ts
@@ -17,7 +17,15 @@ class AuthService {
     }
 
     logout() {
-        TokenService.removeUser();
+        let refreshToken = TokenService.getLocalRefreshToken();
+        return api
+            .post("logout-user", {refreshToken})
+            .then((response) => {
+                if (response.status === 200) {
+                    TokenService.removeUser();
+                }
+                return response
+            });
     }
 }
